feat: add episode type for selected node view
