+++ src/components/App/SideBar/Show/index.tsx
@@ -1,7 +1,6 @@
 import { useMemo, useState } from 'react'
 import styled from 'styled-components'
 import { useGraphData } from '~/components/DataRetriever'
-import PlaylistPlayIcon from '~/components/Icons/PlaylistPlayIcon'
 import { Avatar } from '~/components/common/Avatar'
 import { Flex } from '~/components/common/Flex'
 import { Text } from '~/components/common/Text'
@@ -10,7 +9,7 @@ import { useDataStore } from '~/stores/useDataStore'
 import { NodeExtended } from '~/types'
 import { getSelectedNodeTimestamps } from '~/utils'
 import { colors } from '~/utils/colors'
-import { videoTimeToSeconds } from '~/utils/videoTimetoSeconds'
+import { EpisodePanel } from './EpisodePanel'
 
 export const CREATOR_HEADING_HEIGHT = 240
 
@@ -34,36 +33,6 @@ const Wrapper = styled(Flex)`
   }
 `
 
-type EpisodeProps = {
-  node: NodeExtended
-  onClick: () => void
-}
-
-const EpisodeWrapper = styled(Flex).attrs({})`
-  direction: row;
-  cursor: pointer;
-  color: ${colors.primaryText1};
-
-  .content {
-    margin-left: 16px;
-    align-self: stretch;
-    justify-content: space-between;
-    display: flex;
-    flex-direction: column;
-    width: 100%;
-  }
-
-  .title {
-    margin-top: 16px;
-    display: block;
-  }
-
-  .clipText {
-    font-size: 12px;
-    margin-right: 6px;
-  }
-`
-
 const Header = styled(Flex)`
   flex-direction: column;
   border-bottom: 1px solid #101317;
@@ -84,41 +53,6 @@ const EpisodeHeaderText = styled(Text)`
   display: -webkit-box;
   -webkit-line-clamp: 2;
 `
-
-const EpisodePanel = ({ node, onClick }: EpisodeProps) => {
-  const secs = videoTimeToSeconds(node.timestamp || '')
-  const mins = Math.ceil(secs / 60)
-
-  return (
-    <EpisodeWrapper onClick={onClick} py={12}>
-      <div>
-        <Flex align="flex-start" direction="row" justify="flex-start">
-          <Avatar size={64} src={node?.image_url || 'audio_default.svg'} type="show" />
-          <div className="content">
-            <Flex align="center" direction="row">
-              <TypeBadge type="episode" />
-              {mins > 0 && (
-                <div className="subtitle">
-                  {mins} {mins === 1 ? 'min' : 'mins'}
-                </div>
-              )}
-            </Flex>
-            <Text className="title" color="primaryText1" kind="regular">
-              {node.episode_title}
-            </Text>
-          </div>
-        </Flex>
-
-        <Flex align="center" direction="row" justify="flex-end" pt={4}>
-          <Text className="clipText" color="mainBottomIcons" kind="regular">
-            {node?.children?.length || 0} {node?.children?.length === 1 ? 'Clip' : 'Clips'}
-          </Text>
-          <PlaylistPlayIcon style={{ color: colors.white }} />
-        </Flex>
-      </div>
-    </EpisodeWrapper>
-  )
-}
 
 export const Show = () => {
   const [selectedNode, setSelectedNode] = useDataStore((s) => [s.selectedNode, s.setSelectedNode])
@@ -181,8 +115,8 @@ export const Show = () => {
         </Flex>
       </Header>
 
-      <Flex direction="column" pl={20} pr={20} pt={24}>
-        <Flex pb={20}>
+      <Flex direction="column">
+        <Flex p={24}>
           <Text className="relatedHeader" kind="medium">
             Related Episodes
           </Text>
