+++ src/components/App/SideBar/View/index.tsx
@@ -2,8 +2,8 @@ import { memo } from 'react'
 import { useDataStore, useSelectedNode } from '~/stores/useDataStore'
 import { TextType } from '../../Helper/AskQuestion/Text'
 import { AudioClip } from '../AudioClip'
-import { Creator } from '../Creator'
 import { Data } from '../Data'
+import { Creator } from '../Episode'
 import { Messages } from '../Messages'
 import { Person } from '../Person'
 import { Relevance } from '../Relevance'
